api/requirements.txt
Django>=3.0.8,<3.1
djangorestframework>=3.11.0,<3.12
psycopg2==2.8.3
gunicorn==19.9.0
django-seed>=0.2.2,<1.0
pillow